Create Frontal Obstacle avoidance robot using Fuzzy logic on Arduino
This project proposes Frontal Obstacle avoidance robot using Fuzzy logic on Arduino. Fuzzy logic is the technique in estimating speed and controlling its DC motor. Ultrasonic sensor is used for the detection of the obstacle. As soon as the robot reaches near to the obstacle the speed of its both DC motors varies as is programmed in the Arduino. As soon as the obstacle approaches in front of the robot the speed of its both DC motor adjusts its speed in order to avoid collision with the obstacle. The specific distance limit is defined  below which the robot will change the speed of its both DC motors to avoid the obstacle.